Rework import
When importing from URL, download and process by chunks to speed up the
whole import.

Use expat parser directly (remove the need to access to private elements
of the sax parser).

Display download and processing progress. Progress percentage is only
displayed when importing from URI. (It is technically possible but since
it's not the common use case, it is not supported.)